perform actions in clickedButtonAt... methods instead of didDismiss... methods
I made this change because new UIAlertViews that are shown will send a
alertView:didDismissWithButtonAtIndex: message to alert views already being
shown, with a button index of 0 (the cancel button index). This is particularly
annoying if you don't want an alert view to force an action unexpectedly, like,
say, exiting the application.