Use concrete type in itemWithLabel instead of id
If something else in a project has an object with a label property and that class is ahead of RIButton in the imports, the compiler gives a warning that setLabel: is possibly receiving the wrong type.

Switching to the concrete RIButtonItem type clears this warning. There should be no adverse impact that I'm aware of.